Break out dispatch() and main() into separate files
